+++ backend/IncomeCalculator-backend/src/main/java/pl/twyszomirski/service/IncomeCalculationService.java
@@ -2,10 +2,12 @@ package pl.twyszomirski.service;
 
 import pl.twyszomirski.dto.IncomeCalculationResponseDto;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Tomasz
  */
 public interface IncomeCalculationService {
 
-    IncomeCalculationResponseDto calculateIncome(Float dailyRate, String countryCode) throws NoExchangeRateException;
+    IncomeCalculationResponseDto calculateIncome(BigDecimal dailyRate, String countryCode) throws NoExchangeRateException;
 }
