+++ backend/IncomeCalculator-backend/src/test/java/pl/twyszomirski/util/EntityUtils.java
@@ -2,13 +2,15 @@ package pl.twyszomirski.util;
 
 import pl.twyszomirski.domain.Country;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Tomasz
  */
 public class EntityUtils {
 
     public static Country buildCountry(String countryCode, String currencyCode, String name,
-                                Float taxRate, Long additionalCost){
+                                       BigDecimal taxRate, BigDecimal additionalCost){
         Country country = new Country();
         country.setCountryCode(countryCode);
         country.setName(name);
